+++ projects/command_connectors/src/command_connector/commands/query_git_commit_info.rb
@@ -26,10 +26,6 @@ module Foobara
             def context_type_declaration
               {}
             end
-
-            def context
-              {}
-            end
           end
         end