+++ projects/common/src/runtime_error.rb
@@ -16,7 +16,7 @@ module Foobara
     end
 
     # TODO: why path instead of runtime path?
-    def initialize(message: nil, symbol: nil, context: nil, path: nil)
+    def initialize(message: nil, symbol: nil, context: {}, path: nil)
       args = { message:, symbol:, context:, path: }.compact
       super(**args.merge(category: self.class.category))
     end